Version 1 - Basic Implementation
Simple version of the tracker fully implemented and storing data into an
SQL database.

When a user visits one of the four pages their IP, user agent, visited
page, host name and date/time visited is stored in an SQL database. This
information is stored as soon as the page is loaded. Any HTML page with
the script tag linking to the asp_tracker code will record the data to
the SQL database meaning it is really easy to add the tracker to any web
page, copy and paste 1 line of code!

README.md also updated to reflect new changes.